+++ dhis-2/dhis-services/dhis-service-program-rule/src/main/java/org/hisp/dhis/programrule/api/NotificationEffect.java
@@ -25,20 +25,8 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.programrule.engine;
+package org.hisp.dhis.programrule.api;
 
-import java.util.Arrays;
-
-public enum ValidationActionType {
-  ASSIGN,
-  SETMANDATORYFIELD,
-  SHOWERROR,
-  SHOWWARNING,
-  ERRORONCOMPLETE,
-  WARNINGONCOMPLETE,
-  ERROR;
-
-  public static boolean contains(String value) {
-    return Arrays.stream(values()).anyMatch(v -> v.name().equalsIgnoreCase(value));
-  }
-}
+/** This effect will be used in tracker to send a notification. */
+public record NotificationEffect(
+    String ruleId, String data, String notification, NotificationActionType type) {}
