+++ dhis-2/dhis-services/dhis-service-program-rule/src/main/java/org/hisp/dhis/programrule/api/ValidationActionType.java
@@ -25,7 +25,20 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.programrule.engine;
+package org.hisp.dhis.programrule.api;
 
-public record NotificationEffect(
-    String ruleId, String data, String notification, NotificationActionType type) {}
+import java.util.Arrays;
+
+public enum ValidationActionType {
+  ASSIGN,
+  SETMANDATORYFIELD,
+  SHOWERROR,
+  SHOWWARNING,
+  ERRORONCOMPLETE,
+  WARNINGONCOMPLETE,
+  ERROR;
+
+  public static boolean contains(String value) {
+    return Arrays.stream(values()).anyMatch(v -> v.name().equalsIgnoreCase(value));
+  }
+}
