+++ dhis-2/dhis-services/dhis-service-program-rule/src/main/java/org/hisp/dhis/programrule/engine/DefaultProgramRuleEngine.java
@@ -27,8 +27,6 @@
  */
 package org.hisp.dhis.programrule.engine;
 
-import static org.hisp.dhis.programrule.engine.RuleActionKey.NOTIFICATION;
-
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -44,9 +42,9 @@ import org.hisp.dhis.program.ProgramStage;
 import org.hisp.dhis.programrule.ProgramRule;
 import org.hisp.dhis.programrule.ProgramRuleVariable;
 import org.hisp.dhis.programrule.ProgramRuleVariableService;
+import org.hisp.dhis.programrule.api.RuleEngineEffects;
 import org.hisp.dhis.rules.api.RuleEngine;
 import org.hisp.dhis.rules.api.RuleEngineContext;
-import org.hisp.dhis.rules.models.RuleEffect;
 import org.hisp.dhis.rules.models.RuleEffects;
 import org.hisp.dhis.rules.models.RuleEnrollment;
 import org.hisp.dhis.rules.models.RuleEvent;
@@ -103,76 +101,13 @@ public class DefaultProgramRuleEngine implements ProgramRuleEngine {
             enrollment.getProgram(),
             getRuleEvents(events),
             rules);
-    return map(ruleEffects);
-  }
-
-  private RuleEngineEffects map(List<RuleEffects> ruleEffects) {
-    Map<String, List<ValidationEffect>> enrollmentValidationEffects =
-        ruleEffects.stream()
-            .filter(RuleEffects::isEnrollment)
-            .collect(
-                Collectors.toMap(
-                    RuleEffects::getTrackerObjectUid,
-                    e -> mapValidationEffect(e.getRuleEffects())));
-    Map<String, List<ValidationEffect>> eventValidationEffects =
-        ruleEffects.stream()
-            .filter(RuleEffects::isEvent)
-            .collect(
-                Collectors.toMap(
-                    RuleEffects::getTrackerObjectUid,
-                    e -> mapValidationEffect(e.getRuleEffects())));
-    Map<String, List<NotificationEffect>> enrollmentNotificationEffects =
-        ruleEffects.stream()
-            .filter(RuleEffects::isEnrollment)
-            .collect(
-                Collectors.toMap(
-                    RuleEffects::getTrackerObjectUid,
-                    e -> mapNotificationEffect(e.getRuleEffects())));
-    Map<String, List<NotificationEffect>> eventNotificationEffects =
-        ruleEffects.stream()
-            .filter(RuleEffects::isEvent)
-            .collect(
-                Collectors.toMap(
-                    RuleEffects::getTrackerObjectUid,
-                    e -> mapNotificationEffect(e.getRuleEffects())));
-    return new RuleEngineEffects(
-        enrollmentValidationEffects,
-        eventValidationEffects,
-        enrollmentNotificationEffects,
-        eventNotificationEffects);
-  }
-
-  private List<ValidationEffect> mapValidationEffect(List<RuleEffect> effects) {
-    return effects.stream()
-        .filter(e -> ValidationActionType.contains(e.getRuleAction().getType()))
-        .map(
-            e ->
-                new ValidationEffect(
-                    e.getRuleId(),
-                    e.getData(),
-                    e.getRuleAction().field(),
-                    e.getRuleAction().content(),
-                    ValidationActionType.valueOf(e.getRuleAction().getType())))
-        .toList();
-  }
-
-  private List<NotificationEffect> mapNotificationEffect(List<RuleEffect> effects) {
-    return effects.stream()
-        .filter(e -> NotificationActionType.contains(e.getRuleAction().getType()))
-        .map(
-            e ->
-                new NotificationEffect(
-                    e.getRuleId(),
-                    e.getData(),
-                    e.getRuleAction().getValues().get(NOTIFICATION),
-                    NotificationActionType.valueOf(e.getRuleAction().getType())))
-        .toList();
+    return RuleEngineEffects.fromRuleEffects(ruleEffects);
   }
 
   @Override
   public RuleEngineEffects evaluateProgramEvents(Set<Event> events, Program program) {
     List<ProgramRule> rules = implementableRuleService.getProgramRules(program, null);
-    return map(
+    return RuleEngineEffects.fromRuleEffects(
         evaluateProgramRulesForMultipleTrackerObjects(null, program, getRuleEvents(events), rules));
   }
