+++ dhis-2/dhis-services/dhis-service-program-rule/src/main/java/org/hisp/dhis/programrule/engine/RuleActionImplementer.java
@@ -29,6 +29,7 @@ package org.hisp.dhis.programrule.engine;
 
 import org.hisp.dhis.program.Enrollment;
 import org.hisp.dhis.program.Event;
+import org.hisp.dhis.programrule.api.NotificationEffect;
 
 /**
  * Service is responsible for implementing actions which are generated as a result of Rule-Engine
