+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/imports/programrule/executor/ValidationExecutor.java
@@ -32,7 +32,7 @@ import static org.hisp.dhis.tracker.imports.programrule.ProgramRuleIssue.warning
 
 import java.util.Optional;
 import org.apache.commons.lang3.StringUtils;
-import org.hisp.dhis.programrule.engine.ValidationEffect;
+import org.hisp.dhis.programrule.api.ValidationEffect;
 import org.hisp.dhis.tracker.imports.bundle.TrackerBundle;
 import org.hisp.dhis.tracker.imports.programrule.IssueType;
 import org.hisp.dhis.tracker.imports.programrule.ProgramRuleIssue;
