+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/imports/programrule/executor/event/ShowErrorExecutor.java
@@ -31,7 +31,7 @@ import static org.hisp.dhis.tracker.imports.programrule.IssueType.ERROR;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.hisp.dhis.programrule.engine.ValidationEffect;
+import org.hisp.dhis.programrule.api.ValidationEffect;
 import org.hisp.dhis.tracker.imports.bundle.TrackerBundle;
 import org.hisp.dhis.tracker.imports.domain.Event;
 import org.hisp.dhis.tracker.imports.programrule.IssueType;
